Added wall jumping to parkour. Updated states to transition into wall running without having to press jump btn (ty Julian)
Wall jumping forces can be modified at a per-wall basis if required. Anything over 0 value in forces/vertical/horizontal distance on the climbable surface will automatically override the forces and/or distances for wall running for that particular climbable surface. This is especially useful for walls that are slower/faster (grassy, viney, icy surfaces?) and ones that are harder to jump off of (perhaps a broken or unstable piece of rock?)